world drawing

+++ libs/food.cc
@@ -1,11 +1,11 @@
 #include "food.h"
 
 food::food(){
-	x=0;
-	y=0;
+	x=0.0;
+	y=0.0;
 }
 
-food::food(int _x, int _y){
+food::food(float _x, float _y){
 	x = _x;
 	y = _y;
 }
